### 2019-10-29 dcf4ae5
update setup.py to fix building from source
This commit updates setup.py to add:

    setup(
        ...,
        use_scm_version=...,
        setup_requires=['setuptools_scm'],
        ...,
    )

as recommended by the setuptools-scm project at
https://pypi.org/project/setuptools-scm/.

This commit fixes builds from source which were previously broken due to an ImportError:

    $ git clone https://github.com/dfm/emcee
    $ cd emcee
    $ python setup.py install
    $ cd ~
    $ python
    $ import emcee
    ImportError: No module named emcee_version
