+++ setup.py
@@ -53,6 +53,7 @@ if __name__ == "__main__":
             ),
             "write_to_template": '__version__ = "{version}"\n',
         },
+        setup_requires=['setuptools_scm'],
         author=find_meta("author"),
         author_email=find_meta("email"),
         maintainer=find_meta("author"),
