### 2019-10-29 89ce9eb
adding test for setuptools_scm
