### 2019-10-29 1c2083c
updating install docs
