### 2019-10-29 8ee98e7
package manager docs
