### 2019-10-29 6703e7d
update conda [ci skip]
