### 2019-10-29 750ef63
update pip for source install too [ci skip]
