### 2019-10-30 #1840 @xuanxu
:wave: @dhuppenkothen: would you be willing to review this paper? I think your knowledge of astronomy, statistics and Python makes you a great fit for this submission.
