### 2019-11-02 8004db1
Fixes issue #300
