### 2019-11-02 53cbb84
Applies POLA to resolve issue #313
