### 2019-11-02 #1840 @mattpitkin
Sorry for the delayed response. I'm happy to review this
