### 2019-11-02 #1864 @whedon
[REVIEW]: emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC
**Submitting author:** @dfm (<a href="http://orcid.org/0000-0002-9328-5652">Daniel Foreman-Mackey</a>)
**Repository:** <a href="https://github.com/dfm/emcee" target ="_blank">https://github.com/dfm/emcee</a>
**Version:** v3.0.2
