### 2019-11-02 #1840 @xuanxu
@benjaminrose, @mattpitkin: please head over to #1864 where the review will take place.
