### 2019-11-03 e38081e
fixing nan test
