+++ emcee/tests.py
@@ -32,8 +32,11 @@ def log_unit_sphere_volume(ndim):
         logfactorial = 0.0
         for i in range(1, ndim + 1, 2):
             logfactorial += np.log(i)
-        return (ndim + 1) / 2.0 * np.log(2.0) \
-            + (ndim - 1) / 2.0 * np.log(np.pi) - logfactorial
+        return (
+            (ndim + 1) / 2.0 * np.log(2.0)
+            + (ndim - 1) / 2.0 * np.log(np.pi)
+            - logfactorial
+        )
 
 
 class LogLikeGaussian(object):
@@ -63,7 +66,7 @@ class LogPriorGaussian(object):
 
         if self.cutoff is not None:
             if -dist2 > self.cutoff * self.cutoff / 2.0:
-                return float('-inf')
+                return float("-inf")
             else:
                 return 0.0
         else:
@@ -75,7 +78,6 @@ def ln_flat(x):
 
 
 class Tests:
-
     def setUp(self):
         np.random.seed(42)
 
@@ -87,14 +89,16 @@ class Tests:
         self.N = 3000
 
         self.mean = np.zeros(self.ndim)
-        self.cov = 0.5 - np.random.rand(self.ndim ** 2) \
-            .reshape((self.ndim, self.ndim))
+        self.cov = 0.5 - np.random.rand(self.ndim ** 2).reshape(
+            (self.ndim, self.ndim)
+        )
         self.cov = np.triu(self.cov)
         self.cov += self.cov.T - np.diag(self.cov.diagonal())
         self.cov = np.dot(self.cov, self.cov)
         self.icov = np.linalg.inv(self.cov)
-        self.p0 = [0.1 * np.random.randn(self.ndim)
-                   for i in range(self.nwalkers)]
+        self.p0 = [
+            0.1 * np.random.randn(self.ndim) for i in range(self.nwalkers)
+        ]
         self.truth = np.random.multivariate_normal(self.mean, self.cov, 100000)
 
     def check_sampler(self, N=None, p0=None):
@@ -110,13 +114,15 @@ class Tests:
         assert np.all(self.sampler.acceptance_fraction > 0)
 
         chain = self.sampler.flatchain
-        maxdiff = 10. ** (logprecision)
-        assert np.all((np.mean(chain, axis=0) - self.mean) ** 2 / self.N ** 2
-                      < maxdiff)
-        assert np.all((np.cov(chain, rowvar=0) - self.cov) ** 2 / self.N ** 2
-                      < maxdiff)
+        maxdiff = 10.0 ** (logprecision)
+        assert np.all(
+            (np.mean(chain, axis=0) - self.mean) ** 2 / self.N ** 2 < maxdiff
+        )
+        assert np.all(
+            (np.cov(chain, rowvar=0) - self.cov) ** 2 / self.N ** 2 < maxdiff
+        )
         tau = self.sampler.get_autocorr_time()
-        assert tau.shape == (self.ndim, )
+        assert tau.shape == (self.ndim,)
 
     def check_pt_sampler(self, cutoff, N=None, p0=None):
         if N is None:
@@ -128,52 +134,63 @@ class Tests:
             pass
 
         # Weaker assertions on acceptance fraction
-        assert np.mean(self.sampler.acceptance_fraction) > 0.1, \
-            "acceptance fraction < 0.1"
-        assert np.mean(self.sampler.tswap_acceptance_fraction) > 0.1, \
-            "tswap acceptance frac < 0.1"
+        assert (
+            np.mean(self.sampler.acceptance_fraction) > 0.1
+        ), "acceptance fraction < 0.1"
+        assert (
[…]
