### 2019-11-04 #1864 @xuanxu
@dfm @mattpitkin, @benjaminrose : this is the review thread for the paper. All of our communications will happen here from now on.
