### 2019-11-04 #1864 @benjaminrose
@dfm, it looks via `conda-forge` `emcee` is only at version 2.2.1. Do you have a plan to update `conda-forge` to version 3?
