### 2019-11-04 #1864 @dfm
@benjaminrose: really? It looks like 3.0.0 to me: https://anaconda.org/conda-forge/emcee and that's what I get with:

```bash
conda install -c conda-forge emcee
```

Is that not what you get?
