### 2019-11-04 #1864 @benjaminrose
@dfm After further testing, it appears to be some conflict with Python 3.8 and `conda`. When I switched to an environment running Python 3.7.3 (or install via `pip`) everything works as expected. This is odd, but appears to not be an issue with `emcee`.
