### 2019-11-04 #1864 @dfm
Ah yes - I think that conda-forge is just getting up to speed on 3.8. Should be there soon! Thanks!
