### 2019-11-04 5ed14ce
Fix and test linear dependence check
