### 2019-11-04 3255c0d
Undo mis-blackening
For some reason my editor's "run black" key fails to pick up
pyproject.toml settings.
