+++ src/emcee/ensemble.py
@@ -78,13 +78,19 @@ class EnsembleSampler(object):
     ):
         # Warn about deprecated arguments
         if a is not None:
-            deprecation_warning("The 'a' argument is deprecated, use 'moves' instead")
+            deprecation_warning(
+                "The 'a' argument is deprecated, use 'moves' instead"
+            )
         if threads is not None:
             deprecation_warning("The 'threads' argument is deprecated")
         if runtime_sortingfn is not None:
-            deprecation_warning("The 'runtime_sortingfn' argument is deprecated")
+            deprecation_warning(
+                "The 'runtime_sortingfn' argument is deprecated"
+            )
         if live_dangerously is not None:
-            deprecation_warning("The 'live_dangerously' argument is deprecated")
+            deprecation_warning(
+                "The 'live_dangerously' argument is deprecated"
+            )
 
         # Parse the move schedule
         if moves is None:
@@ -237,7 +243,9 @@ class EnsembleSampler(object):
         state = State(initial_state, copy=True)
         if np.shape(state.coords) != (self.nwalkers, self.ndim):
             raise ValueError("incompatible input dimensions")
-        if (not skip_initial_state_check) and (not walkers_independent(state.coords)):
+        if (not skip_initial_state_check) and (
+            not walkers_independent(state.coords)
+        ):
             raise ValueError(
                 "Initial state has a large condition number. "
                 "Make sure that your walkers are linearly independent for the "
@@ -250,7 +258,8 @@ class EnsembleSampler(object):
         # it's current state.
         if rstate0 is not None:
             deprecation_warning(
-                "The 'rstate0' argument is deprecated, use a 'State' " "instead"
+                "The 'rstate0' argument is deprecated, use a 'State' "
+                "instead"
             )
             state.random_state = rstate0
         self.random_state = state.random_state
@@ -259,7 +268,8 @@ class EnsembleSampler(object):
         # now.
         if log_prob0 is not None:
             deprecation_warning(
-                "The 'log_prob0' argument is deprecated, use a 'State' " "instead"
+                "The 'log_prob0' argument is deprecated, use a 'State' "
+                "instead"
             )
             state.log_prob = log_prob0
         if blobs0 is not None:
@@ -312,7 +322,9 @@ class EnsembleSampler(object):
             map_fn = self.pool.map
         else:
             map_fn = map
-        model = Model(self.log_prob_fn, self.compute_log_prob, map_fn, self._random)
+        model = Model(
+            self.log_prob_fn, self.compute_log_prob, map_fn, self._random
+        )
 
         # Inject the progress bar
         total = iterations * yield_step
@@ -407,7 +419,9 @@ class EnsembleSampler(object):
                 map_func = self.pool.map
             else:
                 map_func = map
-            results = list(map_func(self.log_prob_fn, (p[i] for i in range(len(p)))))
+            results = list(
+                map_func(self.log_prob_fn, (p[i] for i in range(len(p))))
+            )
 
         try:
             log_prob = np.array([float(l[0]) for l in results])
