### 2019-11-04 f86659b
Fix case where longdouble==float
