### 2019-11-04 f1a7e98
adding @aarchiba to JOSS paper
