+++ pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 79
+target-version = ['py35']
 exclude = '''
 /(
     \.eggs
@@ -18,3 +19,11 @@ exclude = '''
   | dist
 )/
 '''
+
+[tool.isort]
+line_length = 79
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+known_third_party = ["h5py", "matplotlib", "numpy", "pkg_resources", "pytest", "scipy", "setuptools"]
