+++ .pre-commit-config.yaml
@@ -22,7 +22,6 @@ repos:
   rev: 19.3b0
   hooks:
   - id: black
-    language_version: python3.6
 
 - repo: https://github.com/dfm/black_nbconvert
   rev: master
