### 2019-11-04 9a97c6f
moving isort config to pyproject.toml
