+++ .isort.cfg
@@ -1,4 +1,0 @@
-[settings]
-line_length=79
-multi_line_output=3
-known_third_party = h5py,matplotlib,numpy,pkg_resources,pytest,scipy,setuptools
