### 2019-11-07 #1864 @benjaminrose
Overall, this is a great software package and I am eager to see it published in JOSS. I do have a few things to discuss before the paper is finalized.

* *Authors* - I fully support the full list of possible authors on https://github.com/dfm/emcee/issues/316. There has been plenty of time for maintainers to request authorship so I am ok with publishing as is or including anyone from that issue.
	* But to be forthright, I work at the same institution as some of the potential co-authors. After reading the JOSS COI policy, I don't think that is an issue nor do I feel unable to make an impartial evaluation of `emcee`. I just wanted to be candid.
* *Performance* - The new parallelization examples are great, but I was unable to get MPI to work on my machine.
* *Docs*
	* *Installation* - Installation, as controllable by this project, is good.
	* *API documentation* - The deprecated parameters are documented as such at runtime. Is there a way to add something like "all other parameters are deprecated" in the online docs?
	* *Community Guidlines* - `Contributing.md` is great. In the documents, there is just a one sentance about using GitHub issues. I think it would be helpful to add a link to both the users mailing list and the `Contribuing.md` file in the offical docs.

I think this is a great project and once we finalize the authors and the best way to inform users about contributing, I will be happy to endurance this paper for publication.
