### 2019-11-11 10aeac0
adding words about deprecations
