### 2019-11-11 23167b9
adding words about contributing to docs
