### 2019-11-11 #1864 @dfm
@benjaminrose: Thanks for this feedback! Do you think that the MPI issues were a problem with emcee/schwimmbad or something else? If the former, please open an issue and let's get it figured out!

I've updated the docs to include [some words about deprecations](https://emcee.readthedocs.io/en/latest/user/sampler/#emcee.EnsembleSampler) and [a bit more info about contributing](https://emcee.readthedocs.io/en/latest/).

As for authors: the current master branch version of the manuscript has all of the authors from the issue, but I'll make one more call for co-authorship requests before finalizing the list if that's ok!

Let me know if there's anything else and thanks again!
