### 2019-11-11 #1864 @benjaminrose
@dfm these changes look great! I think my MPI issues stem mostly from my lack of experience and not an issue with `emcee`.
