### 2019-11-14 #1864 @mattpitkin
Hi @dfm, would you be able to update the references to include more DOIs and a link to the [kombine github page](https://github.com/bfarr/kombine) or [doc page](https://pages.uoregon.edu/bfarr/kombine/)? I think the following are correct:

* Goodman & Weare [10.2140/camcos.2010.5.65](http://dx.doi.org/10.2140/camcos.2010.5.65)
* Ter Braak [10.1007/s11222-006-8769-1](https://doi.org/10.1007/s11222-006-8769-1)
* Ter Braak & Vrught [10.1007/s11222-008-9104-9](https://doi.org/10.1007/s11222-008-9104-9)
