### 2019-11-14 #1864 @mattpitkin
Installation works well, but the conda install gives v3.0.0, whereas pip gives v3.0.1. I'm just curious as to why the [conda-forge feedstock](https://github.com/conda-forge/emcee-feedstock) is not autoupdating to find the latest PyPI version?
