### 2019-11-14 09b4d9c
adding DOIs to JOSS paper
