+++ joss/paper.bib
@@ -14,7 +14,8 @@
 	Date-Modified = {2019-10-17 16:15:45 -0400},
 	Note = {in prep},
 	Title = {kombine: a kernel-density-based, embarrassingly parallel ensemble sampler},
-	Year = 2015}
+	Year = 2015,
+	Url = {https://github.com/bfarr/kombine}}
 
 @article{Ter-Braak:2008,
 	Author = {Ter Braak, Cajo JF and Vrugt, Jasper A},
@@ -26,7 +27,8 @@
 	Publisher = {Springer},
 	Title = {Differential evolution Markov chain with snooker updater and fewer chains},
 	Volume = {18},
-	Year = {2008}}
+	Year = {2008},
+	Doi = {10.1007/s11222-008-9104-9}}
 
 @article{Ter-Braak:2006,
 	Author = {Ter Braak, Cajo JF},
@@ -38,7 +40,8 @@
 	Publisher = {Springer},
 	Title = {A Markov Chain Monte Carlo version of the genetic algorithm Differential Evolution: easy Bayesian computing for real parameter spaces},
 	Volume = {16},
-	Year = {2006}}
+	Year = {2006},
+	Doi = {10.1007/s11222-006-8769-1}}
 
 @article{Speagle:2019,
 	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
@@ -67,7 +70,8 @@
 	Publisher = {Mathematical Sciences Publishers},
 	Title = {Ensemble samplers with affine invariance},
 	Volume = {5},
-	Year = {2010}}
+	Year = {2010},
+	Doi = {10.2140/camcos.2010.5.65}}
 
 @article{Foreman-Mackey:2013,
 	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
