### 2019-11-14 #1864 @mattpitkin
The documentation for emcee is excellent, and is a great resource. All the example work exactly as expected (I know it's expected, but setting the Numpy random seed means I get spookily identical plots when running the example codes). Just a couple of points, one bigger than the other:

* minor point: in the "What about shorter chains?" section it's worth noting that the example requires having autograd installed.
* slightly bigger point: could an example be added that shows using a different Move. Maybe something where the walk move, or differential evolution, given shorter autocorrelation times than the stretch move?
