### 2019-11-14 #1864 @dfm
@mattpitkin: Thanks! I've added those DOIs and a link for kombine. I also merged the pull request to conda-forge so it'll be up to date shortly.
