### 2019-11-14 #1864 @dfm
Those points are also both good. I'll add those too. Thanks!
