### 2019-11-14 dfm/emcee#330 @dfm
Updates for JOSS review
@mattpitkin recommends (ref: openjournals/joss-reviews#1864):

- minor point: in the "What about shorter chains?" section it's worth noting that the example requires having autograd installed.
- slightly bigger point: could an example be added that shows using a different Move. Maybe something where the walk move, or differential evolution, given shorter autocorrelation times than the stretch move?
