### 2019-11-14 #1864 @mattpitkin
Could you add a link to the [CONTRIBUTING.md](https://github.com/dfm/emcee/blob/master/CONTRIBUTING.md) file in the "[How to Use This Guide](https://emcee.readthedocs.io/en/stable/#how-to-use-this-guide)" section of the docs?
