### 2019-11-14 #1864 @dfm
> Could you add a link to the [CONTRIBUTING.md](https://github.com/dfm/emcee/blob/master/CONTRIBUTING.md) file in the "[How to Use This Guide](https://emcee.readthedocs.io/en/stable/#how-to-use-this-guide)" section of the docs?

It is in the introduction of the latest version of the docs: https://emcee.readthedocs.io/en/latest/, but adding it there would also be good.
