### 2019-11-14 019522d
moving comment about contributing to 'how to use this guide section'
