### 2019-11-14 #1864 @mattpitkin
When the Moves example is added I'm happy to sign off the review. All boxes are ticked.
