### 2019-11-14 #1864 @mattpitkin
I was just trying to run the tests (based on `emcee-3.0.2.dev21+gba05c7f`), after installing from source, and I get a range of errors starting:

```
(emcee_pip) matthew@work:~/repositories/emcee$ pytest -v src/emcee/tests
platform linux -- Python 3.7.5, pytest-5.2.2, py-1.8.0, pluggy-0.13.0 -- /home/matthew/miniconda2/envs/emcee_pip/bin/python
cachedir: .pytest_cache
rootdir: /home/matthew/repositories/emcee
collected 0 items / 14 errors

==================================== ERRORS ====================================
___________ ERROR collecting src/emcee/tests/integration/test_de.py ____________
ImportError while importing test module '/home/matthew/repositories/emcee/src/emcee/tests/integration/test_de.py'.
Traceback:
src/emcee/__init__.py:22: in <module>
    from .emcee_version import __version__  # isort:skip
E   ModuleNotFoundError: No module named 'emcee.emcee_version'
```

Have you come across this before?
