### 2019-11-14 #1864 @mattpitkin
It looks like an `emcee_version.py` file is expected in `src/emcee`, but isn't present. Should this be getting created during the setup?
