### 2019-11-14 e3ad451
fixing source install instructions
