### 2019-11-14 #1864 @dfm
Yes - there was a typo in the source install docs! I [just fixed it](https://github.com/dfm/emcee/commit/e3ad45109743b11c1735aa73b2b7110eb984314f), but the basic change is that you need to run:

```bash
pip install -e .
```

 or

```bash
python setup.py develop
```

so that that file is written in place. Good catch!
