### 2019-11-14 02c8a77
adding moves tutorial
