+++ docs/index.rst
@@ -95,6 +95,7 @@ If you have a question about the use of emcee, please post it to `the users list
    tutorials/parallel
    tutorials/autocorr
    tutorials/monitor
+   tutorials/moves
 
 
 License & Attribution
