### 2019-11-14 dfm/emcee#330 @dfm
Done.
