### 2019-11-14 #1864 @dfm
@mattpitkin: I've added [the note about autograd](https://emcee.readthedocs.io/en/latest/tutorials/autocorr/#what-about-shorter-chains) and [a tutorial demonstrating the moves interface](https://emcee.readthedocs.io/en/latest/tutorials/moves/). Let me know if you have any comments or questions!
