### 2019-11-15 b9d6e3e
updating changelog
