### 2019-11-15 #1864 @dfm
Thanks all!!

The new DOI is: [10.5281/zenodo.3543502](https://doi.org/10.5281/zenodo.3543502)

I've also [pushed the update to PyPI](https://pypi.org/project/emcee/) and it'll be on conda-forge in the next few days.
