### 2019-11-15 #1864 @xuanxu
Everything looks ready.
Pinging @openjournals/joss-eics for final processing
