### 2019-11-16 #1864 @labarba
@dfm -- A few small items to address:

1. Originally, emcee was implementation >> an implementation (missing article), or better yet "implemented" (verb form rather than nominalization)
2. Foreman-Mackey, et al. (2013) is missing a journal title
3. Both Ter Braak references need caps protection `{}` in the article title
