### 2019-11-17 fc1068f
tweaks to JOSS paper
