+++ joss/paper.bib
@@ -18,7 +18,7 @@
 	Url = {https://github.com/bfarr/kombine}}
 
 @article{Ter-Braak:2008,
-	Author = {Ter Braak, Cajo JF and Vrugt, Jasper A},
+	Author = {{T}er Braak, Cajo JF and Vrugt, Jasper A},
 	Date-Added = {2019-10-17 16:07:14 -0400},
 	Date-Modified = {2019-10-17 16:15:43 -0400},
 	Journal = {Statistics and Computing},
@@ -31,7 +31,7 @@
 	Doi = {10.1007/s11222-008-9104-9}}
 
 @article{Ter-Braak:2006,
-	Author = {Ter Braak, Cajo JF},
+	Author = {{T}er Braak, Cajo JF},
 	Date-Added = {2019-10-17 16:06:50 -0400},
 	Date-Modified = {2019-10-17 16:15:43 -0400},
 	Journal = {Statistics and Computing},
@@ -82,7 +82,7 @@
 	Date-Modified = {2019-10-17 14:36:36 -0400},
 	Doi = {10.1086/670067},
 	Eprint = {1202.3665},
-	Journal = {\pasp},
+	Journal = {Publications of the Astronomical Society of the Pacific},
 	Keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Statistics - Computation},
 	Month = {Mar},
 	Number = {925},
