### 2019-11-17 b86db1b
tweaks to JOSS paper
