+++ joss/paper.bib
@@ -18,27 +18,27 @@
 	Url = {https://github.com/bfarr/kombine}}
 
 @article{Ter-Braak:2008,
-	Author = {{T}er Braak, Cajo JF and Vrugt, Jasper A},
+	Author = {{ter Braak}, Cajo J.~F. and Vrugt, Jasper A},
 	Date-Added = {2019-10-17 16:07:14 -0400},
 	Date-Modified = {2019-10-17 16:15:43 -0400},
 	Journal = {Statistics and Computing},
 	Number = {4},
 	Pages = {435--446},
 	Publisher = {Springer},
-	Title = {Differential evolution Markov chain with snooker updater and fewer chains},
+	Title = {{Differential evolution Markov chain with snooker updater and fewer chains}},
 	Volume = {18},
 	Year = {2008},
 	Doi = {10.1007/s11222-008-9104-9}}
 
 @article{Ter-Braak:2006,
-	Author = {{T}er Braak, Cajo JF},
+	Author = {{ter Braak}, Cajo J.~F.},
 	Date-Added = {2019-10-17 16:06:50 -0400},
 	Date-Modified = {2019-10-17 16:15:43 -0400},
 	Journal = {Statistics and Computing},
 	Number = {3},
 	Pages = {239--249},
 	Publisher = {Springer},
-	Title = {A Markov Chain Monte Carlo version of the genetic algorithm Differential Evolution: easy Bayesian computing for real parameter spaces},
+	Title = {{A Markov Chain Monte Carlo version of the genetic algorithm Differential Evolution: easy Bayesian computing for real parameter spaces}},
 	Volume = {16},
 	Year = {2006},
 	Doi = {10.1007/s11222-006-8769-1}}
