### 2019-11-17 #1864 @dfm
@labarba: Thanks! Good catches. I've updated the manuscript and BibTeX. Let me know if you have any other suggestions. Should I make another release with these changes?
